refactor(dashboard): split dashboard_mapping into compliance + arch modules
dashboard_mapping.py was 1293 lines, over the 800-line cap in
.claude/rules/coding-style.md. Extract the two largest cohesive concerns
into dedicated modules, behavior-preserving via façade re-export:

- dashboard_compliance.py (464 lines): COMPLIANCE_FRAMEWORKS, the
  compliance-map.py lazy-load + inline COMPLIANCE_CONTROL_MAP fallback,
  _match_prowler_compliance, map_compliance.
- dashboard_arch.py (243 lines): ARCH_DOMAINS, ARCH_DOMAIN_LINKS,
  OWASP_TO_ARCH, map_architecture.

dashboard_mapping.py (now 625 lines) re-exports all moved names with
`# noqa: F401`, so the 9 importer modules need no change — the public
API is identical (verified: all 16 public names resolve through the
façade).

test_dashboard_mapping_pure.py: the fallback-coverage helper
(_reload_with_fallback) and TestFallbackBranch now reload
dashboard_compliance (where the inline fallback now lives) instead of
dashboard_mapping. Same assertions, same intent.

Verification:
- byte-for-byte content of both moved blocks preserved (diff-checked).
- pytest scanner/tests/ → 1368 passed.
- coverage: scanner/lib total 99.12% (>= 99% floor); the 3 affected
  modules each at 100%.
- test_ci_*.py guard suite → 195 passed.
- py_compile clean on all three modules.
